refactor(abstract-utxo): move outputDifference test to abstract-utxo module
Issue: BTC-1450
